git: check-ignore
